requirements.txt
Flask==0.12.1
flask-mysql
azure-storage-blob